Peer info message processing improvements
Factor out the peer_id comparison into a function and use it
in the HOST_WAIT and INIT states. The code is cleaner (although
it is a switch statement inside a switch statement). More
importantly, a peer info message received in HOST_WAIT mode connects
the socket and immediately determines master/slave status. This
expedites the sending of KEY_SWITCH_START, if master.

A HOST_GW in HOST_WAIT should be able to transition directly to the
key exchange phase--similar to INIT, but accepting peer_info fromr
anywhere and not periodically sending them (because it doesn't know
where they should go at this point). There is not need to move to INIT,
then wait for another peer_info to trigger the key exchange.

The result is faster startup, and also *may* avoid a weird, occasional
bug where a HOST_GW, when trying to respond to a new connection in
HOST_WAIT, would get stuck alternating between HOST_WAIT and INIT,
not responding with its own peer_info and not transitioning to
the KEY_SWITCH_START state.

Also: increase the trace level of the "I am key {master,slave}" message.