Add remote nonce persistence
This is the second half of the replay attack prevention:
storing the last received nonce persistently to disk, so
it persists across process invocations. The nonce sending
already periodically stores a "reset point" so it doesn't
look it's sending old packets. With both of these in place
replay attacks are harder.

Also:

- Add a "-V" command line option to print the version.

- Change the port number to 4706.

- When getaddrinfo fails, print the reason (from gai_strerror,
  not the clearest err message generator, unfortunately).

- Update the configuration parameters related to nonce:
  more specific names and new one for the remote nonce.

- Update the man page and README.

- Some indentation cleanups.