Push handling of the tunnel device path into the OS module.
This was a dangling bit of BSD-specific logic in the setup, and it was
clumsy, to boot. Instead of having the full path to the device in the
defaults and then doing the equivalent to "basename" on it, just
specify the name upfront and let the OS-specific module construct
the correct path or otherwise use the name as appropriate.