Make the "remote_network" parameter optional for net-gw.
Allowing this to be unconfigured means that no route is added
when the VPN becomes active and allows admins to do whatever
routing configuration they see fit for the topology.

Run indent(1), the DragonFly version, on changed files.

Centralize the checking of a routing destination in the
utility module and call it from the OS-specific modules.
Initial use is in setting up the net-gw route to the remote
network, but it could be used when configuring the route
in host role as well.