Logging and debug command input enhancements
- Output on stdout, not stderr, and make sure stdout is
  line buffered whether or not it's connected to a terminal.

- Close stdin on EOF, to make sure that it doesn't keep reading
  the EOF indication endlessly.

These enhancements ensure proper operation when running in the
foreground as a child process of boss. We want to make sure that
the output goes to the expected stream and that closing stdin
doesn't cause a loop of transmitting DEBUG_STRING messages to
the peer (which will fail if in host-gw mode and not connected).